[#24] 환경변수 추가, 충돌 오류 해결

See merge request cve-labhub/cve-labhub-backend!26

+++ src/main/java/com/labhub/CveLabhubBack/news/entity/CveNewsMapping.java
@@ -21,7 +21,7 @@ public class CveNewsMapping {
 
     @Id
     @Column(name = "cve_id")
-    private Long cveId;
+    private Integer cveId;
 
     // Composite Key
     @Data
